Merged PR 189673: MSDatePicker with modal support
This PR adds a public `MSDatePicker` class to display and select a date from a calendar style view. It has public APIs to programmatically change the selected date, present the modal, and set a delegate that will be notified when the user has confirmed a date.

It includes additions and changes in support classes for:
* A paged card modal presenter (`MSPageCardPresenterViewController` and supporting classes)
* Multiple styles of obscured backgrounds (`Obscurable`, `MSDimmingView`, and `MSBlurringView`)
* Accessibility navigation support (`AccessibleViewDelegate`)
* A simple dot view to show events on a day (`MSDotView`, unimplemented)
* A number of convenience extensions to help with navigation stacks, date handling, view centering, and more![Screen Shot 2018-10-09 at 6.03.49 PM.png](https://onedrive.visualstudio.com/4dcbf0bc-c3cd-49c8-a7c3-ec1924691d9b/_apis/git/repositories/93ac71ee-b53a-4fc6-a8c4-d46a80d4ca39/pullRequests/189673/attachments/Screen%20Shot%202018-10-09%20at%206.03.49%20PM.png)

Related work items: #640099, #640121, #640470