Merged PR 203007: MSBadgeView with MSBadgeBaseView
This PR introduces MSBadgeView with MSBadgeBaseView to represent a "badge". This badge will be used in an upcoming field to represent selected personas from the MSPersonaListView. The badge can be selected with a tap and then tapped again once in a selected state to reveal more details about the selected person. It includes 3 styles: default, error, and warning with a disabled state as well.

Also included in this PR:
-  A small update to the accessibility for MSAvatarView to become an accessibility element so details about the avatar can be used in VoiceOver when the avatar view is used outside of a MSPersonaCell.
- An update to Physical colors in MSColors to include dark yellow colors for 'warning' and 'light warning'. The previous red colors used for 'warning' and 'light warning' are now named 'error' and 'lightError' respectively.

![Screen Shot 2018-11-14 at 3.56.30 PM.png](https://onedrive.visualstudio.com/4dcbf0bc-c3cd-49c8-a7c3-ec1924691d9b/_apis/git/repositories/93ac71ee-b53a-4fc6-a8c4-d46a80d4ca39/pullRequests/203007/attachments/Screen%20Shot%202018-11-14%20at%203.56.30%20PM.png)

Related work items: #627343