Merged PR 208193: MSBadgeListField for People Picker
This PR introduces the view that contains a list of badges and a textfield that will be used in People Picker named `MSBadgeListField`. The text entered into the field creates a badge that gets appended to a list of badges shown before the view's textfield.

The user can:
- enter text to create badges
- select badges
- drag and drop badges to other badge list fields
- delete badges using the keyboard's delete key
- set a max number of lines for the text field which will append a 'moreBadge' indicating the number of badges not shown if the badges cannot fit within the max number of lines (e.g. "+5")

The next step in the People Picker process will include integration with MSPersonaListView and functionality for this field's entered text to search and view search results in MSPersonaListView where MSPersonaCells can be selected and added to the MSBadgeListView as badges.

![Screen Shot 2018-12-17 at 10.43.46 AM.png](https://onedrive.visualstudio.com/4dcbf0bc-c3cd-49c8-a7c3-ec1924691d9b/_apis/git/repositories/93ac71ee-b53a-4fc6-a8c4-d46a80d4ca39/pullRequests/208193/attachments/Screen%20Shot%202018-12-17%20at%2010.43.46%20AM.png)

Related work items: #627344