Merged PR 251976: Integrate localized files into framework and demo
This PR integrates our localized languages into the Xcode project files for both the framework and the demo.

Every language should now be showing in the demo where appropriate, but we will not be adding localized files for demo-only strings. Localization will show up for any strings that are included in the framework.

This includes fixes for plist syntax errors that were introduced in previous Touchdown PRs.

Related work items: #678557