Merged PR 252376: Fix localization integration
This PR creates a strings file for every supported language in the demo.

The previous localization PR only worked during testing because of caching in the build folder, and would stop supporting the expected languages when the build folder was cleaned. Supporting a language requires at least one localized file, so we are adding a LaunchScreen.strings for every language.

Related work items: #678557