Merged PR 260439: Fix DatePickerController dismiss bug
Added a new optional delegate method to `MSDateTimePickerDelegate` that allows a consumer to turn off modal auto-dismiss when the done button is pressed. This allows the user to validate or do asynchronous work before either dismissing themselves via `MSDateTimePicker.dismiss()`, or not dismiss altogether, effectively canceling a dismiss. If the delegate method is not implemented, auto-dismiss when done is pressed defaults to on.

Also added an example of this to the demo with a switch to turn on said validation.