Add maxPointSize to MSLabel to specify a largest DynamicType size (#4)
* Add maxPointSize to MSLabel

For some UI elements, it doesn't make sense to scale indefinitely with the DynamicType size. This change allows us to scale up to a maximum point size, then maintain it even if the user selects a larger DynamicType size. This could be used for non-core text in a view, for example, the author's name and title.