Merged PR 269114: Add a static lib build target
Some clients (namely Office) require a static lib instead of a framework for consuming fabric-ios. Add a new target that builds all the same sources.

Note: The framework build target is taking advantage of the umbrella header to import UIKit into every Swift file. Since the static library isn't doing so, we need to have each file import the proper Apple framework to build as expected.

There will be a follow up pull request to switch the NuGet publishing pipeline over to packaging the static lib and the resources bundle instead of the framework.