Merged PR 269480: Rename static library target
Rename static library target

We initially wanted to keep the static library name the same as the framework, but this causes an issue where both the static library and framework produce identically named modules. This causes a conflict in building the test app where it compiles against the static library module but links in the framework.

Let's uniquely name the static library to OfficeUIFabricLib to avoid any ambiguity.