Merged PR 288800: Fix MSCollectionViewCell sizing issue when using estimatedItemSize
Recent update to MSTableViewCell `intrinsicContentSize` and` sizeThatFits()` returning `preferredWidth` causes MSCollectionViewCell to ignore the width passed into `estimatedItemSize`. Fixed by implementing `preferredLayoutAttributesFitting()` which enables an update of layout attributes of the cell using the attributes provided by the collectionViewLayout object.

Can also just change `intrinsicContentSize` and `sizeThatFits()` in MSCollectionViewCell to not return `preferredWidth` for the returned size's width to fix this issue but preserving the preferredWidth for both of these might be useful going forward so opting for `preferredLayoutAttributesFitting()` solution for now.

![Screen Shot 2019-07-02 at 5.56.10 PM.png](https://onedrive.visualstudio.com/4dcbf0bc-c3cd-49c8-a7c3-ec1924691d9b/_apis/git/repositories/93ac71ee-b53a-4fc6-a8c4-d46a80d4ca39/pullRequests/288800/attachments/Screen%20Shot%202019-07-02%20at%205.56.10%20PM.png)

Related work items: #744668