Merged PR 310671: Make MSTableViewCell's customAccessoryView getter public
Office file cells use a custom accessory view, which consist of several accessory items (pin, share, file actions, error). Given that cells are constantly reused and not created every time, for performance we want to also reuse/reconfigure (if needed) the custom accessory view we created the cell with, instead of creating a new one every time and reloading the same images. Making the getter public will allow this. It will also make it easy for the accessory view and the cell to be configured in different code places/classes, each of which could be unaware of cell or accessory view data model objects (that's the case for Office).