Merged PR 318118: New MScrollView
- New `MScrollView`
- `UIResponder` extension: new `firstResponder`
- Demo uses new `MScrollView`

Main reason for this work it to make scroll view used in our demo automatically adjust content insets when keyboard is shown/hidden. This is required for People Picker demo. The rest of functionality is there to support keyboard focus correct - we don't use this yet but I decided to keep the code. The source came from the code I've written while in my old team.