Merged PR 319614: iOS 13 QuickType Keyboard Infinite Loop Avoidance
This PR addresses an issue in iOS 13 in which the QuickType keyboard violates expectations around delegate calls.

Specifically, if the textField.text value is changed programmatically, the UITextFieldDelegate call shouldChangeCharactersInRange() should NOT be invoked. This behavior is documented and is relied on for the current flow of data in the MSSearchBar's implementation of shouldChangeCharactersInRange()

However, if the user is typing using iOS 13's new QuickType keyboard, this expectation is violated and the result is an infinite loop in which the delegate method is called when the textfield.text property is changed programmatically.

To avoid this issue, the delegate method is no longer used to respond to changes in the textfield. Instead, we introduce a UIControl target/action.

Link to the OneDrive PR that addressed this issue, also containing a larger description/links to the Apple Feedback Assistant ticket, and other info: https://onedrive.visualstudio.com/SkyDrive/_git/OneDrive.iOS/pullrequest/312125