Merged PR 355016: [Build Settings] Reference info plist in unit test xcconfig
## Problem
We have a warning when building that our unit test target is "Skipping code signing because the target does not have an Info.plist file and one is not being generated automatically".

## Fix
We do have an info plist for this target, we just completely forgot to hook it up when switching from pbxproj settings to xcconfig settings. Just add the INFOPLIST_FILE build setting and point to the appropriate one.

## Client Impact
None, unit test target only change

## Validation
Build and run tests, expect to see the build succeed with no warnings at all.