Merged PR 356513: [Avatar View] Fix initials extraction algorithm and add more tests
## Problem
Our initials algorithm wasn't properly trimming whitespace and zero spacing characters. It was trimming them when determining if a word was valid, but then using the untrimmed strings to grab the initials.

Separately, using the first unicode scalar as the initial wasn't quite accurate, as a single Character can be represented by multiple unicode scalars (like unicode diacritics which are their own scalar).

## Fix
Switch from checking an individual unicode scalar for validity and move to checking a single character. Verify that each of the unicode scalars contained within that character are valid as well as that the character as a whole is convertible to Mac OS Roman.

Add unit tests that capture the  multiple unicode scalar case as well as ensure we're handling whitespace, zero width spaces, and newlines as expected.

## Client Impact
Outward facing API remains the same, but if clients were passing in names with whitespace, newlines, zero width spaces, or complex characters represented by multiple unicode scalars, expect more correct initials extraction.

## Validation
The passing of existing and additional unit tests are the validation for this change.

Related work items: #3738895