Merged PR 300705: MSTableViewCell - Support for metadata UI (title/subtitle/footer leading and trailing accessory views)
Adds leading and trailing label accessory views inline with the title, subtitle, and footer labels (named `titleLeadingAccessoryView`, `titleTrailingAccessoryView`, `subtitleLeadingAccessoryView`, `subtitleTrailingAccessoryView`, `footerLeadingAccessoryView`, `footerTrailingAccessoryView` respectively) designed to show metadata about the cell - could include text, icons showing status such as sharing, completion, or a progress indicator for example.

The leading accessory view is positioned on the leading edge of the text area while the trailing accessory view takes up the available space of the text area on the trailing edge of the label. If no text is given for a specific label the label accessory views will influence the `layoutType` of the cell to make sure there's an appropriate amount of space vertically available to show the accessory view. If the label accessory view width is too large for the available space the label text will be truncated at a smaller width to make room for the accessory view.

Each label accessory view will be vertically centered with the label, however no assumptions or restrictions are made for the height of the accessory view as it will be up to the developer to make sure the accessory view will visually fit with the other views within the height of the cell.

- Updates to cell demos showing a few different variations of leading and trailing label accessory views including left aligned, full width, and right aligned content for trailing accessory views specifically.
- Some general clean up and renaming also included in `MSTableViewCell`.
     - Renamed many properties / method names from using "left/right" to "leading/trailing" respectively.

 ![Screen Shot 2019-10-08 at 1.26.14 PM.png](https://onedrive.visualstudio.com/4dcbf0bc-c3cd-49c8-a7c3-ec1924691d9b/_apis/git/repositories/93ac71ee-b53a-4fc6-a8c4-d46a80d4ca39/pullRequests/300705/attachments/Screen%20Shot%202019-10-08%20at%201.26.14%20PM.png)
![Screen Shot 2019-10-08 at 1.26.32 PM.png](https://onedrive.visualstudio.com/4dcbf0bc-c3cd-49c8-a7c3-ec1924691d9b/_apis/git/repositories/93ac71ee-b53a-4fc6-a8c4-d46a80d4ca39/pullRequests/300705/attachments/Screen%20Shot%202019-10-08%20at%201.26.32%20PM.png)
![Screen Shot 2019-10-08 at 1.29.38 PM.png](https://onedrive.visualstudio.com/4dcbf0bc-c3cd-49c8-a7c3-ec1924691d9b/_apis/git/repositories/93ac71ee-b53a-4fc6-a8c4-d46a80d4ca39/pullRequests/300705/attachments/Screen%20Shot%202019-10-08%20at%201.29.38%20PM.png)
![Screen Shot 2019-10-24 at 11.33.32 AM.png](https://onedrive.visualstudio.com/4dcbf0bc-c3cd-49c8-a7c3-ec1924691d9b/_apis/git/repositories/93ac71ee-b53a-4fc6-a8c4-d46a80d4ca39/pullRequests/300705/attachments/Screen%20Shot%202019-10-24%20at%2011.33.32%20AM.png)
![Screen Shot 2019-08-01 at 1.27.33 PM.png](https://onedrive.visualstudio.com/4dcbf0bc-c3cd-49c8-a7c3-ec1924691d9b/_apis/git/repositories/93ac71ee-b53a-4fc6-a8c4-d46a80d4ca39/pullRequests/300705/attachments/Screen%20Shot%202019-08-01%20at%201.27.33%20PM ...

Related work items: #687937, #762641