Merged PR 362354: [Framework] Add a product bundle identifier
## Problem
Apple rejected an app from the app store containing this framework since it didn't have a bundle identifier.

## Fix
Add a bundle identifier. Choose to add it to the xcconfig instead of to the info plist directly

## Client Impact
Apps should pass store validation

## Validation
We don't have a test app that we can submit for validation assurance just yet, but we should probably come up with a way to validate that a simple app using the control library works eventually.

Instead, verify that in the Xcode UI under General -> Identity that the value shows up in the bundle identifier field.