Merged PR 342323: [NuGet Builds] Build library for distribution for NuGet builds
By default, if you're importing a swift module compiled with a different version of the Swift compiler, you'll get an error that the contents can't be imported.

If you use the BUILD_LIBRARY_FOR_DISTRIBUTION flag when building, an extra file (.swiftinterface) is made with human readable text that allows for interop between different versions of the Swift compiler. Add that to our NuGet publish phase so that we are more resilient to compiler differences.