Merged PR 365606: [Compliance] Add [test] and [security-analysis] tags for build compliance
## Problem
The OE system for detecting functional/security tests does not consider Apple platforms at all, so it doesn't detect that we're testing using Xcode.

## Fix
Add the tags to force it to know we're testing via Xcode.

## Client Impact
None

## Validation
Ensure the names in the PR build phase have the tags as expected.