Merged PR 344387: MSTableViewCell - Take custom accessory view's height into consideration for cell height
Update height and preferredWidth calculations as well as layout to support the cell resizing its height based on the height of the custom accessory view if it is a label.

Min-width of the text area added as the min of either half of the text area's leading offset to the custom accessory view's trailing offset or preferred width of the text area itself if it is smaller. The custom accessory view can then take the remaining space and resize to multiple lines if needed.

If the custom accessory view is a label with multiple lines and is greater in height the text area will center itself vertically in the cell.

![Screen Shot 2019-11-12 at 3.55.57 PM.png](https://onedrive.visualstudio.com/4dcbf0bc-c3cd-49c8-a7c3-ec1924691d9b/_apis/git/repositories/93ac71ee-b53a-4fc6-a8c4-d46a80d4ca39/pullRequests/344387/attachments/Screen%20Shot%202019-11-12%20at%203.55.57%20PM.png)

Related work items: #795135