Merged PR 375993: [DatePicker] Make the color of the date picker selection customizable
## Problem
In some cases, a client might want to theme the date picker color to match an in-app theme instead of directly following the system accent color

## Fix
In our CalendarDayButton, add a `color` var that we'll use to override the default system background color. Pipe that color down from the `DatePickerController` to the `CalendarDayButton` so that our clients can directly set it on the controller.

Be sure to grab the cgColor in `updateLayer` to ensure we react to appearance changes as expected.

## Client Impact
Adds a new public API for setting the color. If a client doesn't use this new API, behavior remains unchanged.

## Validation
Add test app code to exercise the color change. Ensure that the colors update live when changing the color via the new test app color picker.

![Screen Shot 2019-12-03 at 3.22.43 PM.png](https://office.visualstudio.com/bc6e0ce9-86c2-4c5a-aece-d99afc758a73/_apis/git/repositories/e1a6d148-62f6-4cbb-9c64-7fc1667e3c25/pullRequests/375993/attachments/Screen%20Shot%202019-12-03%20at%203.22.43%20PM.png)

[Video of it in action](https://office.visualstudio.com/bc6e0ce9-86c2-4c5a-aece-d99afc758a73/_apis/git/repositories/e1a6d148-62f6-4cbb-9c64-7fc1667e3c25/pullRequests/375993/attachments/Screen%20Recording%202019-12-03%20at%203.15.58%20PM.mov)

Related work items: #3554152