Merged PR 380955: [Nuget] Zip frameworks before packing
## Problem
The NuGet pack command hard resolves all symlinks and doesn't support packing up symlinks. This causes our frameworks to have copies of files where they really should just have symlinks.

## Fix
Add a bash script to our publish template that copies the build output we care about into a folder structure that mimics what we had before and zips it. Zipping it with the symlinks option before packaging it preserves the symlinks through the nuget packing/unpacking process.
/.ado/templates/fabric-macos-nuget-publish-job.yml

In our nuspec file, package up the zip file in the "Archives" folder. By convention, when importing this nuget package, any zip files in the archives folder will be expanded into the root of the nuget folder.
/OfficeUIFabric.nuspec

## Client Impact
Clients consuming fabric-macos via nuget package (Office in DevMain) will have a framework that works with codesigning.

## Validation
Run a build and publish a new version of the package based on the the new logic. Confirm locally that importing in DevMain results in the proper build output and the framework codesigns as expected.
https://dev.azure.com/office/ISS/_packaging?_a=package&feed=Office&package=Microsoft.Office.UIFabric.macOS&protocolType=NuGet&version=1.2019.1211-v1