Merged PR 356387: MSDrawerController - Support for scroll-to-resize
Now, when `contentScrollView` property is set and drawer is presented `.up`, it will expand and collapse as result of scrolling in scroll view (and scrolling will be blocked until drawer is done expanding/collapsing). This is similar to panel behavior in Google Maps, Apple Maps (iOS 13) and others.

To try this out in the demo:
1) Open MSDrawerController demo
2) Tap the most bottom button "Show always as slideover, resizable"
3) Try to scroll down the content - drawer will expand and block scrolling until it's fully expanded
4) Try to scroll up the content to zero offset and beyond - drawer will collapse and block scrolling

Related work items: #675018