Merged PR 357893: Add custom link opening handler
On Outlook we would link to be able to open links in the default browser for Outlook. To do this, add a delegate so the main app can handle opening the link rather than opening the link in the default browser. Make this optional so the link will open in the default browser by default.

Related work items: #811076