Merged PR 358297: MSTableViewHeaderFooterView - Removed limit to text storage to avoid possible text truncation
When MSTableViewHeaderFooterView is reused `textContainer.maximumNumberOfLines` is reset to 1 line and so when large text is assigned only 1 line of it gets stored, unless `titleNumberOfLines` is set to correct (larger) value BEFORE assigning a text. This creates fragile API. Removed limit to text storage - the view height will limit how many lines of text will be shown.

- MSTableViewHeaderFooterView:
  - no need to limit text storage to `titleNumberOfLines` - this creates problems when view is reused and `titleNumberOfLines` is set after text is assigned
- TableViewHeaderFooterSampleData:
  - added sample for truncated header

Related work items: #811344