Merged PR 385069: [DatePicker] Deliver formatted strings to CalendarDayButton
This is the first PR for #3554153 to pipe correct data to the calendar buttons. The next PR will contain the updates to the calendar buttons to actually show this data (will display up to two day identifiers instead of just one).

## Problem
CalendarDayButtons currently store a date and generate the string representation using their own date formatter. This uses the user's calendar, which could differ from the calendar supplied when the date picker was created. For secondary calendar support, the buttons will need to display two identifiers if a secondary calendar is supplied.

## Fix
Move formatting logic to the controller, supply day buttons with a struct containing the necessary data to display a day.

## Client Impact
Mostly no impact. Day formatting will now use the supplied calendar, but this results in no change in most cases because the day buttons only show the day component.

## Validation
Manually tested using different locales, updated unit tests.

Related work items: #3554153