Merged PR 387924: [DatePicker] Remove DatePickerScrollView
## Problem
I noticed we are subclassing `NSScrollView` in the date picker, but not using anything it gives us. It was also causing a scrolling bug because of the way manual scrolling is disabled (#3864827). **~~`NSScrollView` uses `NSClipView` internally and adds some scrolling functionality on top of it. In this case it makes sense to just use `NSClipView` directly, because it gives us everything we need and prevents any future manual scrolling bugs.~~**

## Fix
**~~Remove `DatePickerScrollView`, add `DatePickerClipView` with similar functionality.~~**
After a few updates it turns out `NSClipView` is also not necessary, so `NSView` is now used for the clipping.

## Client Impact
Fixes #3864827. Prevents future manual scrolling bugs.

## Validation
Sanity checks in the test app. Tested in and out of NSMenu and in RTL.

Related work items: #3864827