Merged PR 317534: MSPeoplePicker
Implements `MSPeoplePicker` which is used to select personas by providing a list of `suggestedPersonas` in a `MSPersonaListView` under a badge field consisting of a text field which filters the `suggestedPersonas` and a list of badges which displays the list of `pickedPersonas`.

The text field can filter the `availablePersonas` or the delegate method `getSuggestedPersonasFromText` can be implemented to provide the list of `suggestedPersonas`. Once picked the personas will be displayed as an `MSBadgeView` which can be interacted with and removed.

- Updates to `MSBadgeField` included to make it work better with `MSPeoplePicker`. This includes the delegate not being required since `MSPeoplePicker` when set up with sufficient data will function as intended without most delegate methods implemented, making them optional.
- Small updates to `MSBadgeField` demo and clean up in `MSPersonaListView` demo.

![Screen Shot 2019-09-11 at 5.37.20 PM.png](https://onedrive.visualstudio.com/4dcbf0bc-c3cd-49c8-a7c3-ec1924691d9b/_apis/git/repositories/93ac71ee-b53a-4fc6-a8c4-d46a80d4ca39/pullRequests/317534/attachments/Screen%20Shot%202019-09-11%20at%205.37.20%20PM.png)

Related work items: #663382