Merged PR 388542: [Documentation] Add an AvatarView documentation page
## Problem
We don't have consumption documentation for our controls in place, making it harder for consumers to understand what our controls provide.

## Fix
Add a new "Controls" folder to the existing root docs folder that includes markdown documentation for the AvatarView. This is patterned off of the same information provided at uifabric.azurewebsites.net documentation so that when we are ready to be added to there we have all the right content.

Also add a template file to make adding new controls to the documentation easier.

## Client Impact
Documentation only, no code impact.

## Validation
View the documentation with the ADO markdown preview to ensure it looks reasonable.