Merged PR 394057: [DatePicker] Add a flag to toggle auto selection when paging
## Problem
Automatic selection when paging is hardcoded. Not every usage scenario wants this behavior.

## Fix
Add a flag to `DatePickerController` to control whether auto selection should occur when paging.

Note:
Because paging used to be coupled to selection, some functions in `DatePickerView` used the `selectedDate` to populate the header and the calendar views. I changed these to use the visible range, which is now the only source of truth for what's actually displayed.

## Client Impact
Default behavior stays the same, so there should be no impact.

## Validation
Added a checkbox to the test app. Manually validated functionality in the test app. Also added a mock delegate to check that selection events come as expected.

![Screen Shot 2020-01-13 at 2.25.38 PM.png](https://office.visualstudio.com/bc6e0ce9-86c2-4c5a-aece-d99afc758a73/_apis/git/repositories/e1a6d148-62f6-4cbb-9c64-7fc1667e3c25/pullRequests/394057/attachments/Screen%20Shot%202020-01-13%20at%202.25.38%20PM%20(2).png)

Related work items: #3873439