Merged PR 400726: [DatePicker] Fix incorrect month shown for some initial dates
## Problem
When the date picker is initialized with a date that's not in the current month, the day grid shows the current month instead of the month of the initial date.
This bug was introduced with the latest date picker change, when we started relying on the visible range instead of the selected date to generate the displayed months.

## Fix
Remove constructor that initially sets the visibleRange to today. This also ensures that it is set before super.init.

## Client Impact
To-Do app is already pulling the fabric-macos version with the bug, so this will fix it for them.

## Validation
Manually tested in the test app and the To-Do app. Changed the test date used in the test app to be a date in the past to cover this type of bug.

Related work items: #3909069