Merged PR 401462: [DatePicker] Remove delegate notifications when date setter is used
## Problem
We need to be able to set a date on `DatePickerController` without triggering a delegate notification.

## Fix
~~Add a setter with a `shouldNotifyDelegate` flag.~~
Disable delegate notifications for programmatic date changes.

## Client Impact
There should be no impact to current clients. This will be used in the Excel integration.

## Validation
~~Added unit tests which check for delegate calls in multiple scenarios.~~
Added a unit test to check that the delegate is not notified.

Related work items: #3908282