Merged PR 400517: [DatePicker] Add hasEdgePadding flag to datePickerController
## Problem
The date picker will frequently be displayed with margins around it. In cases where it's presented in NSPopover, there is no clean way to achieve this without wrapping it in another view controller. It makes sense to include this on the fluent side.

## Fix
~~Add `edgeInsets` property to `DatePickerController`, which modifies the same property of the underlying `DatePickerView`.
Setting the property in `DatePickerView` updates the edge constraints.~~
Added the `hasEdgePadding` flag to toggle padding.

## Client Impact
No impact to existing clients, the API change is backward-compatible. For future clients, it will allow for cleaner integration.

## Validation
Added new fields to the test app and updated the layout. Manually tested in the test app and in the ToDo app by choosing a variety of edge insets.
![Screen Shot 2020-01-24 at 1.48.55 PM.png](https://office.visualstudio.com/bc6e0ce9-86c2-4c5a-aece-d99afc758a73/_apis/git/repositories/e1a6d148-62f6-4cbb-9c64-7fc1667e3c25/pullRequests/400517/attachments/Screen%20Shot%202020-01-24%20at%201.48.55%20PM.png)

Related work items: #3908279