Merged PR 403229: [DatePicker] Add hasTextField flag to DatePickerController
## Problem
We need to be able to hide the text date picker from the date picker view.

## Fix
Added the `hasTextDatePicker` flag to `DatePickerView` and `DatePickerController`.
Setting the flag just sets the `isHidden` property of the `textDatePicker` to the inverse of the new value.
Getting the flag returns the inverse of the `isHidden` property.

## Client Impact
No impact as the default stays the same (`hasTextDatePicker == true`)

## Validation
1. Added a checkbox tied to the flag to the test app.
2. In the test app, checked that disabling the text date picker works well together with and without edge padding and in NSMenu.
3. Checked that the text date picker disappears from the accessibility hierarchy when it's hidden.
4. Added a unit test which performs a few sanity checks on the expected change in fitting size of the date picker view.

Related work items: #3875413