Merged PR 407226: [DatePicker] Slim down CalendarDayButton
##\#1 in a series of performance PRs

Date picker first start is slow. I noticed that most of the time is spent in `init` and `layoutSubtreeIfNeeded`, so those two are what I'm timing and comparing when making these changes. `layoutSubtreeIfNeeded` performance mostly correlates with the number of NSViews in the hierarchy. Because we have many CalendarDayButtons (42 per CalendarView), even a single unnecessary NSView in this class has a large impact on overall performance.

## Problem
`CalendarDayButton` uses two unnecessary container views to position the underlying labels. It also inits these labels when it doesn't need to (we only need two labels if have a secondary calendar). I also noticed the upper label in dual mode is always the same as the centered label in normal mode, so we can reuse it.

## Fix
1. Get rid of the two container views and position the labels directly.
2. Only init the upper label initially. Init the lower label on demand if we have a string for it.
3. Reuse the upper label to be the centered label in normal mode by switching between two centerY constraints.

## Client Impact
The date picker is noticeably faster to open in scenarios when it's not cached. There is some perf improvement in the cases where it's cached as well, because of the reduced number of NSViews.
No visual changes and no API changes.

Both `init` and `layoutSubtreeIfNeeded` duration down by ~40%

## Validation
1. Manually tested in NSPopover, standalone, and in NSMenu.
2. Verified no visual change in both single and dual calendar mode.

Related work items: #3940962