Merged PR 407900: [DatePicker] Remove the left/right buffer of CalendarViews
##\#3 in a series of performance PRs
\#1 is [here](https://office.visualstudio.com/ISS/_git/fabric-macos/pullrequest/407226?_a=overview),
\#2 is [here](https://office.visualstudio.com/ISS/_git/fabric-macos/pullrequest/407719?_a=overview).

## Problem
Currently we keep 3 `CalendarViews` in memory at all times, with 2 of them as a buffer on the left and right to ensure responsive paging. I made this design choice with the assumption that creating a new `CalendarView` on demand when paging would be too slow. It turns out that this is not that slow, but does become significant when we are creating 3 of them, which is what happens when the date picker is first init-ed and presented. The 2 buffered `CalendarViews` also stay in the view hierarchy, significantly reducing AutoLayout performance because of the number of `CalendarViewButton` instances (126).

## Fix
Only keep one `CalendarView` in the view hierarchy. Create new ones on demand when the next/previous button is pressed.
All the important stuff is in the `DatePickerView.animateInDirection` method.

## Client Impact
No visual or API change. Significantly faster Init and AutoLayout performance. This will affect both the Excel and ToDo app integrations.
AutoLayout duration down by ~60%
Init duration down by ~50%.

## Validation
Manual validation in the test app. Checked different flag combos, NSPopover, NSMenu, LTR, RTL, manual date input through the text field, ToDo integration.

Related work items: #3940962