Merged PR 410647: [DatePicker] Fix text field cursor reset after inputting distant dates
## Problem
Inputting any date that's more than one month away initiates a refresh of `DatePickerView`. This also means that the text date picker gets the current calendar from `DatePickerController`, which in turn causes the cursor inside the text field to reset. This isn't apparent when we just change the month, as that is the first input segment anyways, but it is easily reproduced by just inputing a year, for example 2024.

## Fix
To fix this we simply check if the currently used calendar differs from what the datasource provides. If it doesn't, we don't set it again and hence don't trigger the cursor reset.

## Client Impact
Makes the user experience much better for users who use the text field. Currently the text field is only used in the ToDo app.

## Validation
Manual checks to see that the bug does not happen anymore. Tried custom calendars to make sure that the text date picker is correctly set up with the custom calendar the first time.

Related work items: #3873384