Merged PR 384177: Popup Menu - Support for description
Previously, when only `title` was provided in the `headerItem`, we showed popup menu header as 1-line cell. This PR changes this to show a special "description UI" (smaller, horizontally centered label) instead per Mobile Toolkit design. Also multiline text is supported in this case. When both `title` and `subtitle` are provided for header, then the behavior stays the same: 2-line cell will be used to present header.
| Single line description | Multiline description |
|-|-|
| ![image.png](https://onedrive.visualstudio.com/4dcbf0bc-c3cd-49c8-a7c3-ec1924691d9b/_apis/git/repositories/93ac71ee-b53a-4fc6-a8c4-d46a80d4ca39/pullRequests/384177/attachments/image.png) | ![image.png](https://onedrive.visualstudio.com/4dcbf0bc-c3cd-49c8-a7c3-ec1924691d9b/_apis/git/repositories/93ac71ee-b53a-4fc6-a8c4-d46a80d4ca39/pullRequests/384177/attachments/image%20(2).png) |

Related work items: #707038