Merged PR 412784: [DatePicker] Respect system reduce motion setting
## Problem
The date picker component does not respect the system reduce motion setting.

## Fix
Disable the calendar sliding animation when the reduce motion setting is enabled.

## Client Impact
The control is more accessible. This is explained in more detail in [this](https://aka.ms/reduce-motion) spec.

## Validation
Verified that paging works as expected with the setting both on an off.

Related work items: #3956149