Merged PR 423974: [DatePicker] Fix default button title getting drawn on CalendarDayButtons on High Sierra
## Problem
On High Sierra, if the title property of our button is set to a non-empty string, `NSButtonTextField` gets added to the view hierarchy with that string. This title is set to "Button" by default, and results in broken drawing on High Sierra.

## Why does it not repro on 10.14+
`NSButton` drawing just seems to be handled differently since Mojave. On 10.14+ having `wantsUpdateLayer == true` is sufficient to completely override NSButton drawing, regardless of the `title` value. Even with `wantsUpdateLayer == false`, while the title now gets drawn, `NSButtonTextField` does not show up in the hierarchy. If anyone has a better knowledge of what changed please let me know.

## Fix
Set the title to an empty string in `CalendarDayButton` init function.

## Client Impact
Fixes the bug which impacts both of our Excel and ToDo clients on High Sierra.

## Validation
Reproduced the bug on a High Sierra vm and verified that the fix works. Also verified that the `NSButtonTextField` does not appear in the view hierarchy at all if the title string is empty (good for perf reasons).

![Screen Shot 2020-03-04 at 11.00.30 AM.png](https://office.visualstudio.com/bc6e0ce9-86c2-4c5a-aece-d99afc758a73/_apis/git/repositories/e1a6d148-62f6-4cbb-9c64-7fc1667e3c25/pullRequests/423974/attachments/Screen%20Shot%202020-03-04%20at%2011.00.30%20AM.png)

Related work items: #3997788