Merged PR 432745: [License] Add MIT License comment to top of all files
## Problem
Before pushing to Open Source, we should have the headers of each file indicate the code is licensed under the MIT license, just like we do currently in fabric-ios

## Fix
Add the license statement to each file in the project.

## Client Impact
None, comment only.

## Validation
Manually run through each file in the repository, ensure if it is a text-style file it contains the notice at the top.