1. Add frameImage public property for MSFAvatarView 2. Add imageForAccessoryButtonColor public property for MSFTableViewHeaderFooterView
1. frameImage can be set so that the Avatar has a 3 pixel frame (~broder) based on this image. This image will be used to create a color, which eventually is set as the background color of the avatar frame view.

Why not expose a color property instead of an image? Because the avatar view size is variable and not public. If a client creates a color with an image, the image is fixed in size and the color patter won't be updated when avatar size changes occur.
Why not expose a frame/border view property instead of an image? Because the Avatar View uses really creates 2 avatar views (big and small) for contract/extract. One view can't be shared by 2 hierarchies, and can't be deep copied either.

2. imageForAccessoryButtonColor can be set so that the color of the text in the accessory button is based on an image, overriding the accessory button style's predefined color.

Why not expose a color property instead of an image? Same as for avatar, the text length is variable and not public.
Why not expose a way to override the accessory view entirely? This would be a large refactor and we'd loose some button properties we still want to keep Fluent consistent, like text font and size.

Validation:

Create Demo for both. Played with using the avatar frame in navigation view controller (expanded, contracted).