Modify MSFAvatarView to support custom border image (#74)
Add customBorderImage public property for MSFAvatarView

customBorderImage can be set so that the Avatar has a 3 pixel frame (~broder) based on this image. This image will be used to create a color, which eventually is set as the background color of the avatar frame view.

Why not expose a color property instead of an image? Because the avatar view size is variable and not public. If a client creates a color with an image, the image is fixed in size and the color patter won't be updated when avatar size changes occur.
Why not expose a frame/border view property instead of an image? Because the Avatar View uses really creates 2 avatar views (big and small) for contract/extract. One view can't be shared by 2 hierarchies, and can't be deep copied either.


Validation:

Create Demo for both. Played with using the avatar frame in navigation view controller (expanded, contracted).

* Address CR comments. Will work on table view header footer's accessory button in a separate larger change.

* Address Pull Request comments, use just one border view for AvatarView